feat: added player entity

+++ src/main/java/com/rsh/fcl/dto/PlayRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record PlayRequest(
-    @NotNull @Positive Integer batsman,
-    @NotNull @Positive Integer bowler,
+    @NotNull @Positive Long batsman,
+    @NotNull @Positive Long bowler,
     @NotNull Integer outcome) {
 }
